exercises/os-and-networks/pyproject.toml
[tool.poetry]
name = "task_for_srvhub"
version = "0.1.0"
description = ""
authors = ["Yury Borodich <xkxixnxgx@yandex.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.7.8"
beautifulsoup4 = "4.9.1"
certifi = "2020.6.20"
chardet = "3.0.4"
click = "7.1.2"
Flask = "1.1.2"
Flask-Login = "0.5.0"
Flask-SQLAlchemy = "2.4.4"
Flask-WTF = "0.14.3"
idna = "2.10"
itsdangerous = "1.1.0"
Jinja2 = "2.11.3"
MarkupSafe = "1.1.1"
requests = "2.26.0"
SQLAlchemy = "1.3.19"
urllib3 = "1.26.5"
Werkzeug = "1.0.1"
WTForms = "2.3.3"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
